+++ src/ChaosTools.jl
@@ -33,6 +33,9 @@ include("periodic.jl")
 # Chaos Detection:
 include("chaos_detection.jl")
 
+# Ugly methods that shouldn't exist:
+include("ugliness.jl")
+
 # Visualization routines:
 # using Requires
 # @require PyPlot include("visualizations.jl")